requirement.txt
streamlit
openpyxl
requests
pillow  # For image handling (added for safety in uploads)